Clean up channel stuff a bit
